refactor(constants): rank enum 패키지 수정
